requirements.txt
# Used only for local dev and CI
pytest
moto[dynamodb]==5.1.18
pytest-mock
# Include the layer's requirements for local linting/testing
-r ./serverless/db_layer/requirements.txt

testcontainers[postgres]
testcontainers[localstack]
boto3
psycopg2-binary

azure.functions
azure-cosmos